created the shapeCreator function
