+++ src/App.js
@@ -1,54 +1,13 @@
 import "./App.css";
 import * as THREE from "three";
 import { useEffect } from "react";
-import { OrbitControls } from "three/addons/controls/OrbitControls.js";
-import Stats from "three/examples/jsm/libs/stats.module.js";
+import CubeCreator from "./ThreeJS/CubeCreator.jsx";
 
 function App() {
   useEffect(() => {
-    const scene = new THREE.Scene();
-    const camera = new THREE.PerspectiveCamera(
-      50,
-      window.innerWidth / window.innerHeight,
-      1,
-      1000
-    );
-    camera.position.z = 96;
-
-    const canvas = document.getElementById("myThreeJsCanvas");
-    const renderer = new THREE.WebGL1Renderer({
-      canvas,
-      antialias: true,
-    });
-    renderer.setSize(window.innerWidth, window.innerHeight);
-    document.body.appendChild(renderer.domElement);
-
-    const ambientLight = new THREE.AmbientLight(0xffffff, 0.5);
-    ambientLight.castShadow = true;
-    scene.add(ambientLight);
-
-    const spotLight = new THREE.SpotLight(0xffffff, 1);
-    spotLight.castShadow = true;
-    spotLight.position.set(0, 64, 32);
-    scene.add(spotLight);
-
-    const boxGeometry = new THREE.BoxGeometry(16, 16, 16);
-    const boxMaterial = new THREE.MeshNormalMaterial();
-    const boxMesh = new THREE.Mesh(boxGeometry, boxMaterial);
-    scene.add(boxMesh);
-
-    const controls = new OrbitControls(camera, renderer.domElement);
-    const stats = Stats();
-    document.body.appendChild(stats.dom);
-
-    const animate = () => {
-      stats.update();
-      controls.update();
-      renderer.render(scene, camera);
-      window.requestAnimationFrame(animate);
-    };
-
-    animate();
+    // const shapeSelector = new THREE.BoxGeometry(16, 16, 16);
+    const shapeSelector = new THREE.TorusGeometry(15, 3, 16, 100);
+    CubeCreator("myThreeJsCanvas", shapeSelector);
   }, []);
 
   return (
